move some syscall stuff back to netlink.go
